requirements.txt
Django==2.2.7
djangorestframework==3.10.3
gunicorn==19.9.0
Pillow==6.2.1
pytz==2019.3
requests==2.12.4
sqlparse==0.3.0
urllib3==1.19.1
setuptools==41.6.0
django-heroku==0.3.1
psycopg2==2.8.4
django-rest-auth==0.9.5
dj-database-url==0.5.0
whitenoise==4.1.3
fpdf==1.7.2